# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "prot-cellar",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/organisms/resolve/{tax_id}": {
      "get": {
        "tags": [
          "organisms"
        ],
        "summary": "Resolve Organism",
        "operationId": "resolve_organism_api_v1_organisms_resolve__tax_id__get",
        "parameters": [
          {
            "name": "tax_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tax Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganismResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organisms": {
      "get": {
        "tags": [
          "organisms"
        ],
        "summary": "List Organisms",
        "operationId": "list_organisms_api_v1_organisms_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "name": "rank",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rank"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_OrganismResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "organisms"
        ],
        "summary": "Create Organism",
        "operationId": "create_organism_api_v1_organisms_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganismBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganismResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organisms/{organism_id}": {
      "get": {
        "tags": [
          "organisms"
        ],
        "summary": "Get Organism",
        "operationId": "get_organism_api_v1_organisms__organism_id__get",
        "parameters": [
          {
            "name": "organism_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Organism Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganismResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "organisms"
        ],
        "summary": "Update Organism",
        "operationId": "update_organism_api_v1_organisms__organism_id__patch",
        "parameters": [
          {
            "name": "organism_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Organism Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganismBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganismResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organisms/bulk": {
      "post": {
        "tags": [
          "organisms"
        ],
        "summary": "Bulk Upsert Organisms",
        "operationId": "bulk_upsert_organisms_api_v1_organisms_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/protcellar__interface__routes__organisms__BulkUpsertBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkUpsertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkSummaryResponse": {
        "properties": {
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "updated": {
            "type": "integer",
            "title": "Updated"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          }
        },
        "type": "object",
        "required": [
          "created",
          "updated",
          "skipped",
          "failed"
        ],
        "title": "BulkSummaryResponse"
      },
...
      "BulkUpsertResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/ItemResultResponse"
            },
            "type": "array",
            "title": "Results"
          },
          "summary": {
            "$ref": "#/components/schemas/BulkSummaryResponse"
          }
        },
        "type": "object",
        "required": [
          "results",
          "summary"
        ],
        "title": "BulkUpsertResponse"
      },
...
      "CreateOrganismBody": {
        "properties": {
          "ncbi_tax_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ncbi Tax Id"
          },
          "rank": {
            "type": "string",
            "title": "Rank"
          },
          "scientific_name": {
            "type": "string",
            "title": "Scientific Name"
          },
          "source": {
            "$ref": "#/components/schemas/OrganismSource",
            "default": "ncbi"
          },
          "division": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Division"
          },
          "source_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Version"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "scientific_name"
        ],
        "title": "CreateOrganismBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemResultResponse": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "index",
          "status"
        ],
        "title": "ItemResultResponse"
      },
...
      "NameClass": {
        "type": "string",
        "enum": [
          "scientific_name",
          "common_name",
          "genbank_common_name",
          "synonym",
          "authority",
          "equivalent_name",
          "acronym",
          "blast_name",
          "uniprot_mnemonic"
        ],
        "title": "NameClass"
      },
      "OrganismNameResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "name_class": {
            "$ref": "#/components/schemas/NameClass"
          },
          "unique_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unique Name"
          },
          "is_preferred": {
            "type": "boolean",
            "title": "Is Preferred"
          }
        },
        "type": "object",
        "required": [
          "name",
          "name_class",
          "is_preferred"
        ],
        "title": "OrganismNameResponse"
      },
      "OrganismResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ncbi_tax_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ncbi Tax Id"
          },
          "ncbi_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ncbi Url"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "rank": {
            "type": "string",
            "title": "Rank"
          },
          "scientific_name": {
            "type": "string",
            "title": "Scientific Name"
          },
          "division": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Division"
          },
          "reference_strain_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Strain Id"
          },
          "is_merged": {
            "type": "boolean",
            "title": "Is Merged"
          },
          "merged_into_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merged Into Id"
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted"
          },
          "source": {
            "$ref": "#/components/schemas/OrganismSource"
          },
          "source_release": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Release"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "names": {
            "items": {
              "$ref": "#/components/schemas/OrganismNameResponse"
            },
            "type": "array",
            "title": "Names"
          }
        },
        "type": "object",
        "required": [
          "id",
          "rank",
          "scientific_name",
          "is_merged",
          "is_deleted",
          "source",
          "version",
          "names"
        ],
        "title": "OrganismResponse"
      },
      "OrganismSource": {
        "type": "string",
        "enum": [
          "ncbi",
          "gtdb",
          "local"
        ],
        "title": "OrganismSource"
      },
...
      "PaginatedResponse_OrganismResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrganismResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PaginatedResponse[OrganismResponse]"
      },
...
      "UpdateOrganismBody": {
        "properties": {
          "scientific_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scientific Name"
          },
          "rank": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "division": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Division"
          },
          "reference_strain_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Strain Id"
          },
          "source_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UpdateOrganismBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "protcellar__interface__routes__organisms__BulkRecordBody": {
        "properties": {
          "ncbi_tax_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ncbi Tax Id"
          },
          "rank": {
            "type": "string",
            "title": "Rank"
          },
          "scientific_name": {
            "type": "string",
            "title": "Scientific Name"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "source_release": {
            "type": "string",
            "title": "Source Release"
          },
          "source_record_id": {
            "type": "string",
            "title": "Source Record Id"
          },
          "source_record_checksum": {
            "type": "string",
            "title": "Source Record Checksum"
          },
          "division": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Division"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "scientific_name",
          "source",
          "source_release",
          "source_record_id",
          "source_record_checksum"
        ],
        "title": "BulkRecordBody"
      },
      "protcellar__interface__routes__organisms__BulkUpsertBody": {
        "properties": {
          "records": {
            "items": {
              "$ref": "#/components/schemas/protcellar__interface__routes__organisms__BulkRecordBody"
            },
            "type": "array",
            "title": "Records"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "records"
        ],
        "title": "BulkUpsertBody"
      },
...
    }
  }
}